WIP: begin to test Trust-Store with configuration file
NOTES: To make the whitelist directory portable, i.e. get the right
path configured, we might need to do some work on the configuration
file before the test-set gets going. The parts that write out
certificate files mightn't change because they simply need to write to
the right place which is where the Trust-Store will be reading from.